+++ Source/TwoD.h
@@ -12,7 +12,7 @@ protected:
 	// My warning to you: comment as you code. 
 	// Because a year later I have no idea where 767 and 1367 came from
 	// And I'm to scared to change them now.
-	const float SCREENTOP = 0, SCREENBOTTOM = 767,
+	const double SCREENTOP = 0, SCREENBOTTOM = 767,
 		SCREENLEFT = 0, SCREENRIGHT = 1367;
 
 	// Prepares OpenGL draw in 2D
